backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
sqlalchemy
bcrypt
PyJWT
